requirements.txt
Pillow==10.4.0
python-telegram-bot==21.6
PyYAML==6.0.2
torch==2.4.0
torchvision==0.19.0
torchaudio==2.4.0
transformers==4.44.2
watchdog==3.0.0
numpy==1.23.5
pandas==1.5.3
matplotlib==3.7.0
scikit-learn==1.2.2
tqdm==4.66.5
einops==0.7.0
safetensors==0.4.4
